server/requirements.txt
alabaster==0.7.13
anyio==3.6.2
awscrt==0.14.0
Babel==2.12.1
bcrypt==4.0.1
beautifulsoup4==4.11.2
black==23.3.0
boto3==1.24.31
botocore==1.27.96
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cmake==3.25.2
cryptography==39.0.2
databind==1.5.3
databind.core==1.5.3
databind.json==1.5.3
Deprecated==1.2.13
docspec==2.1.2
docspec-python==2.1.2
docstring-parser==0.11
docutils==0.19
ecdsa==0.18.0
fastapi==0.95.0
flake8==6.0.0
furo==2022.12.7
h11==0.14.0
httptools==0.5.0
idna==3.4
imagesize==1.4.1
isort==5.10.1
Jinja2==3.1.2
jmespath==1.0.1
mangum==0.17.0
MarkupSafe==2.0.1
mccabe==0.7.0
mypy-boto3-dynamodb==1.24.0
mypy-boto3-ec2==1.24.0
mypy-boto3-s3==1.24.0
mypy-boto3-secretsmanager==1.24.0
mypy-boto3-ses==1.24.0
mypy-extensions==0.4.3
nr.util==0.8.12
packaging==23.0
parsimonious==0.10.0
passlib==1.7.4
pathspec==0.10.3
platformdirs==2.6.2
protobuf==4.21.12
psycopg2==2.9.5
psycopg2-binary==2.9.6
pyasn1==0.4.8
pyclean==2.2.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.7
pydoc-markdown==4.6.4
pyflakes==3.0.1
Pygments==2.14.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.5
pywatchman==1.4.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
rsa==4.9
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.4
Sphinx==5.3.0
sphinx-basic-ng==1.0.0b1
sphinx-js==3.2.1
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==2.0.8
sqlalchemy-cockroachdb==2.0.0
starlette==0.26.1
tomli==2.0.1
tomli_w==1.0.0
typing_extensions==4.5.0
tzdata==2022.6
urllib3==1.26.14
uvicorn==0.21.1
uvloop==0.17.0
watchdog==3.0.0
watchfiles==0.19.0
websockets==11.0
wrapt==1.15.0
yapf==0.32.0
